DP-1219 Syntax overhaul for `origo datasets cp`
Overhaul the syntax of the `origo datasets cp` command. Four different
URI formats are now supported when specifying the dataset to upload to
or download from:

- `ds:{dataset_id}`
- `ds:{dataset_id}/{version}`
- `ds:{dataset_id}/{version}/latest`
- `ds:{dataset_id}/{version}/{edition}`

The two first forms will automatically create new editions on upload.

Support for the `ds://` prefix and support for providing positional
versions and editions has been removed.