T#223 Clarify Maskinporten environment input
Clarify that we're after the Maskinporten environment in every
`pubreg` sub-command. Every sub-command is turned into a wizard,
avoiding having to specify the environment on the command line (as a
nice side-effect, users no longer have to copy/paste client IDs and
key IDs between commands).

We might look into adding non-interactive support for all these
commands later, if there is demand for it.